+++ apps/dotcom/src/utils/createAssetFromUrl.ts
@@ -9,77 +9,44 @@ interface ResponseBody {
 }
 
 export async function createAssetFromUrl({ url }: { type: 'url'; url: string }): Promise<TLAsset> {
+	const urlHash = getHashForString(url)
 	try {
 		// First, try to get the meta data from our endpoint
-		const meta = (await (
-			await fetch(BOOKMARK_ENDPOINT, {
-				method: 'POST',
-				headers: {
-					'Content-Type': 'application/json',
-				},
-				body: JSON.stringify({
-					url,
-				}),
-			})
-		).json()) as ResponseBody
+		const fetchUrl =
+			BOOKMARK_ENDPOINT +
+			'?' +
+			new URLSearchParams({
+				url,
+			}).toString()
+
+		const meta = (await (await fetch(fetchUrl)).json()) as ResponseBody | null
 
 		return {
-			id: AssetRecordType.createId(getHashForString(url)),
+			id: AssetRecordType.createId(urlHash),
 			typeName: 'asset',
 			type: 'bookmark',
 			props: {
 				src: url,
-				description: meta.description ?? '',
-				image: meta.image ?? '',
-				favicon: meta.favicon ?? '',
-				title: meta.title ?? '',
+				description: meta?.description ?? '',
+				image: meta?.image ?? '',
+				favicon: meta?.favicon ?? '',
+				title: meta?.title ?? '',
 			},
 			meta: {},
 		}
 	} catch (error) {
-		// Otherwise, fallback to fetching data from the url
-
-		let meta: { image: string; favicon: string; title: string; description: string }
-
-		try {
-			const resp = await fetch(url, {
-				method: 'GET',
-				mode: 'no-cors',
-			})
-			const html = await resp.text()
-			const doc = new DOMParser().parseFromString(html, 'text/html')
-			meta = {
-				image: doc.head.querySelector('meta[property="og:image"]')?.getAttribute('content') ?? '',
-				favicon:
-					doc.head.querySelector('link[rel="apple-touch-icon"]')?.getAttribute('href') ??
-					doc.head.querySelector('link[rel="icon"]')?.getAttribute('href') ??
-					'',
-				title: doc.head.querySelector('meta[property="og:title"]')?.getAttribute('content') ?? '',
-				description:
-					doc.head.querySelector('meta[property="og:description"]')?.getAttribute('content') ?? '',
-			}
-			if (meta.image.startsWith('/')) {
-				meta.image = new URL(meta.image, url).href
-			}
-			if (meta.favicon.startsWith('/')) {
-				meta.favicon = new URL(meta.favicon, url).href
-			}
-		} catch (error) {
-			console.error(error)
-			meta = { image: '', favicon: '', title: '', description: '' }
-		}
-
-		// Create the bookmark asset from the meta
+		// Otherwise, fallback to a blank bookmark
+		console.error(error)
 		return {
-			id: AssetRecordType.createId(getHashForString(url)),
+			id: AssetRecordType.createId(urlHash),
 			typeName: 'asset',
 			type: 'bookmark',
 			props: {
 				src: url,
-				image: meta.image,
-				favicon: meta.favicon,
-				title: meta.title,
-				description: meta.description,
+				description: '',
+				image: '',
+				favicon: '',
+				title: '',
 			},
 			meta: {},
 		}