+++ packages/tldraw/src/lib/shapes/shared/useAsset.ts
@@ -1,11 +1,18 @@
-import { TLAssetId, useEditor, useValue } from '@tldraw/editor'
-import { useEffect, useState } from 'react'
+import { TLAssetId, TLShapeId, useEditor, useValue } from '@tldraw/editor'
+import { useEffect, useRef, useState } from 'react'
 
 /** @internal */
-export function useAsset(assetId: TLAssetId | null, width: number) {
+export function useAsset(shapeId: TLShapeId, assetId: TLAssetId | null, width: number) {
 	const editor = useEditor()
 	const [url, setUrl] = useState<string | null>(null)
 	const asset = assetId ? editor.getAsset(assetId) : null
+	const culledShapes = editor.getCulledShapes()
+	const isCulled = culledShapes.has(shapeId)
+	const didAlreadyResolve = useRef(false)
+
+	useEffect(() => {
+		if (url) didAlreadyResolve.current = true
+	}, [url])
 
 	const shapeScale = asset && 'w' in asset.props ? width / asset.props.w : 1
 	// We debounce the zoom level to reduce the number of times we fetch a new image and,
@@ -16,19 +23,24 @@ export function useAsset(assetId: TLAssetId | null, width: number) {
 	])
 
 	useEffect(() => {
+		if (isCulled) return
+
 		let isCancelled = false
-		const timer = editor.timers.setTimeout(async () => {
-			const resolvedUrl = await editor.resolveAssetUrl(assetId, {
-				screenScale,
-			})
-			if (!isCancelled) setUrl(resolvedUrl)
-		}, 500)
+		const timer = editor.timers.setTimeout(
+			async () => {
+				const resolvedUrl = await editor.resolveAssetUrl(assetId, {
+					screenScale,
+				})
+				if (!isCancelled) setUrl(resolvedUrl)
+			},
+			didAlreadyResolve.current ? 500 : 0
+		)
 
 		return () => {
 			clearTimeout(timer)
 			isCancelled = true
 		}
-	}, [assetId, asset?.props.src, screenScale, editor])
+	}, [assetId, asset?.props.src, isCulled, screenScale, editor])
 
 	return { asset, url }
 }