+++ scripts/publish-vscode-extension.ts
@@ -1,11 +1,12 @@
 import { existsSync, readFileSync, writeFileSync } from 'fs'
 import path from 'path'
+import { parse } from 'semver'
 import { exec } from './lib/exec'
 import { makeEnv } from './lib/makeEnv'
+import { nicelog } from './lib/nicelog'
 
 // VSCE_PAT needs to be set. It is used by the vsce publish command.
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const env = makeEnv(['VSCE_PAT'])
+const env = makeEnv(['VSCE_PAT', 'TLDRAW_ENV'])
 
 const EXTENSION_DIR = 'apps/vscode/extension'
 
@@ -19,9 +20,13 @@ async function updateExtensionVersion() {
 	if (!version) {
 		throw new Error('Could not get the version of the published extension.')
 	}
-	const versionSplit = version.split('.')
-	versionSplit[2] = Number(versionSplit[2]) + 1
-	const nextVersion = versionSplit.join('.')
+	const semVer = parse(version)
+	if (!semVer) {
+		throw new Error('Could not parse the published version.')
+	}
+	const release = env.TLDRAW_ENV === 'production' ? 'minor' : 'patch'
+	const nextVersion = semVer.inc(release).version
+	nicelog(`Updating extension version from ${version} to ${nextVersion}`)
 
 	const packageJsonPath = path.join(EXTENSION_DIR, 'package.json')
 	if (!existsSync(packageJsonPath)) {
@@ -34,8 +39,19 @@ async function updateExtensionVersion() {
 }
 
 async function packageAndPublish() {
-	await exec('yarn', ['package'], { pwd: EXTENSION_DIR })
-	await exec('yarn', ['publish'], { pwd: EXTENSION_DIR })
+	await exec('yarn', ['lazy', 'run', 'build', '--filter=packages/*'])
+	switch (env.TLDRAW_ENV) {
+		case 'production':
+			await exec('yarn', ['package'], { pwd: EXTENSION_DIR })
+			await exec('yarn', ['publish'], { pwd: EXTENSION_DIR })
+			return
+		case 'staging':
+			await exec('yarn', ['package', '--pre-release'], { pwd: EXTENSION_DIR })
+			await exec('yarn', ['publish', '--pre-release'], { pwd: EXTENSION_DIR })
+			return
+		default:
+			throw new Error('Workflow triggered from a branch other than main or production.')
+	}
 }
 
 async function main() {